+++ app/models/event.rb
@@ -10,6 +10,7 @@ class Event < ApplicationRecord
 
   # Associations
   has_many :attendances, dependent: :destroy
+  has_many :user_events, dependent: :destroy
 
   private