+++ config/routes.rb
@@ -15,5 +15,6 @@ Rails.application.routes.draw do
   # HighVoltage
   get '/dashboard' => 'dashboards#index', id: 'dashboard'
 
-  resources :attendance, only: [:new, :index]
+  resources :attendance, only: %i(new index)
+  resources :events, only: %i(show edit destroy)
 end