+++ spec/rails_helper.rb
@@ -65,6 +65,8 @@ RSpec.configure do |config|
   # arbitrary gems may also be filtered via:
   # config.filter_gems_from_backtrace("gem name")
 
+  config.include Devise::Test::IntegrationHelpers, type: :system
+
   config.before(:each, type: :system) do
     driven_by :rack_test
   end