+++ spec/support/dashboard_helpers.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+# dashboard helper function
+def get_event_rows(page)
+  table = page.find(:css, 'table#events')
+  table.all(:css, 'tbody/tr')
+end